fix: 모든 명령어에 arguments가 필요했던 오류 수정
